Add support for Global Rate Limits
https://repose.atlassian.net/wiki/display/REPOSE/Rate+Limiting+filter#RateLimitingfilter-Globalratelimits

This allows you to optionally turn on Global Rate Limits very similarly
to how it is done for Rate Limit Groups elsewhere in this cookbook,
albeit slightly different due to how Global Rate Limits are implemented
in Repose.

This requires Repose >= 6.1.0.3, but if youre not there yet you have
bigger problems. :trollface: